backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose==3.3.0
celery==5.4.0
redis==5.1.1
python-multipart==0.0.12
pytest==8.3.3
httpx==0.27.2
email-validator==2.2.0
